debug

+++ core/representantetecnico/views/bodega/delete.py
@@ -1,11 +1,13 @@
+from django.contrib import messages
 from django.contrib.auth.mixins import LoginRequiredMixin
-from django.http import JsonResponse
+from django.http import JsonResponse, HttpResponseRedirect
 from django.urls import reverse_lazy
 from django.views.generic import DeleteView
 
 from app.settings import LOGIN_REDIRECT_URL
 from core.base.mixins import ValidatePermissionRequiredMixin
 from core.bodega.models import Bodega
+from core.representantetecnico.models import Stock
 
 
 class BodegaDeleteView(LoginRequiredMixin, ValidatePermissionRequiredMixin, DeleteView):
@@ -17,11 +19,18 @@ class BodegaDeleteView(LoginRequiredMixin, ValidatePermissionRequiredMixin, Dele
 
     def dispatch(self, request, *args, **kwargs):
         self.object = self.get_object()
+        if Stock.objects.filter(bodega_id=self.object.id).exists():
+            messages.error(request, 'Bodega ya posee stock registrado')
+            messages.error(request, 'No es posible su eliminación')
+            messages.error(request, 'Pongase en contacto con el administrador del sistema')
+            return HttpResponseRedirect(self.success_url)
         return super().dispatch(request, *args, **kwargs)
 
     def post(self, request, *args, **kwargs):
         data = {}
         try:
+            if Stock.objects.filter(bodega_id=self.object.id).exists():
+                raise Exception("No es posible eliminar este registro")
             self.object.delete()
         except Exception as e:
             data['error'] = str(e)
